pyproject.toml
[project]
name = "buttplug"
version = "1.0.0"
description = "Buttplug Intimate Hardware Control Library"
readme = "README.md"
license = "BSD-3-Clause"
requires-python = ">=3.10"
authors = [
    { name = "Nonpolynomial Labs", email = "kyle@nonpolynomial.com" }
]
keywords = ["buttplug", "haptics", "teledildonics", "hardware", "iot"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: BSD License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Typing :: Typed",
]
dependencies = [
    "websockets>=12.0",
    "pydantic>=2.0",
    "pyserial>=3.5",
    "pyyaml>=6.0",
    "platformio>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "mypy>=1.0",
    "ruff>=0.1",
]

[project.urls]
Homepage = "https://buttplug.io"
Repository = "https://github.com/buttplugio/buttplug-py"
Documentation = "https://docs.buttplug.io"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/buttplug"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_configs = true
plugins = ["pydantic.mypy"]

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.14.14",
]
